DorkHunter/requirements.txt
googlesearch-python
requests